backend/requirements.txt
Django>=4.2,<5.0

# API
djangorestframework>=3.14,<3.15

# Feed sync
paramiko>=3.3,<4.0

# Optional background jobs
celery>=5.3,<6.0
